requirements.txt
colorama
comtypes
concurrent_log_handler
pywin32
PyYAML
tabulate
streamlit
plotly
pandas
numpy
google-generativeai
python-dotenv